Change vector type from Vec<u64> to Vec<u8> in example
    Updated the type of the `vec` field to Vec<u8> for better compatibility or efficiency. This change ensures consistency with the expected use case of the example.

+++ examples/vectors.rs
@@ -13,7 +13,7 @@ pub struct TestData {
     // Regular numeric field that won't be hashed
     pub age: u8,
     #[hash]
-    pub vec: Vec<u64>,
+    pub vec: Vec<u8>,
 }
 
 fn main() {
